requirements.txt
einops==0.3.0
matplotlib==3.1.3
mmseg==1.3.0
mmsegmentation==0.18.0
numpy==1.18.1
opencv_python==4.4.0.44
Pillow==9.4.0
scipy==1.4.1
thop==0.0.31.post2005241907
timm==0.4.12
torch==1.9.0+cu102
torchvision==0.10.0+cu102
